screen-capture/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx>=0.25.0
mss>=9.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
